Removed file system checking holdover code. Added SQL persistence layer. Updated data model structure. Moved code over to async. Updated CLI arguments to match database types Changed repository layer to async trait implementation. Removed unnecessary error types.

+++ src/arguments.rs
@@ -32,8 +32,8 @@ pub enum Commands {
     },
 
     #[clap(about = "Removes a task")]
-    Remove { number: usize },
+    Remove { number: i32 },
 
     #[clap(about = "Completes a task")]
-    Complete { number: usize },
+    Complete { number: i32 },
 }
